backend/pyproject.toml
[project]
name = "admin-backend"
version = "0.1.0"
description = "后台管理系统后端服务（FastAPI）"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aerich>=0.7,<1.0",
  "aiofiles>=23,<25",
  "aiomysql>=0.2,<1.0",
  "asyncpg>=0.29,<1.0",
  "fastapi>=0.125,<1.0",
  "passlib[bcrypt]>=1.7.4,<2.0",
  "pydantic-settings>=2,<3",
  "pydantic>=2,<3",
  "python-jose[cryptography]>=3.3,<4.0",
  "python-multipart>=0.0.9,<1.0",
  "redis>=5,<6",
  "tortoise-orm>=0.25,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
]

[dependency-groups]
dev = [
  "httpx>=0.27,<1.0",
  "mypy>=1.10,<2.0",
  "pytest-asyncio>=0.23,<1.0",
  "pytest>=8,<9",
  "ruff>=0.5,<1.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["migrations"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.aerich]
tortoise_orm = "app.core.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./"
